requirements.txt
Django==1.7.6
django-autofixture==0.9.2
djangorestframework==3.1.0
envvars==0.3.0
Pygments==2.0.2
